Added patrol and pause states
The enemy now patrols around the map through random position vectors, but the vectors are not necessarily random. They are manipulated to make the enemies get closer to the player over time.

Added a patrol timer that allows us to change how long until the next patrol path is chosen.

Added a "Huh???" timer as a pause timer to allow for realistic pauses in a character, between states.

Added a Stun timer to allow for any stun mechanics we implement.