Added nav region global group and fiddled with large_room_CGBoxes.tscn nav region #23
I messed with the baking values to see if I can get a larger navigation area, a lower navigation height (to prevent a region being baked on the top of the walls and a floating region).

Also attempted to debug the problem with enemy agents hitting the walls, to no avail. Navigation Regions are harddd.